Updated controllers to INHERIT from baseController.
