+++ app/controllers/parties_controller.rb
@@ -1,4 +1,4 @@
-class PartiesController < ApplicationController
+class PartiesController < Users::BaseController
   before_action :set_movie, only: [:create]
   def new
     session[:movie] = { title: params[:title],
