+++ app/controllers/users/base_controller.rb
@@ -1,0 +1,11 @@
+class Users::BaseController < ApplicationController
+  before_action :require_login
+
+  private
+  def require_login
+    unless current_user
+      flash[:error] = "You must be logged in to access this section"
+      redirect_to root_path
+    end
+  end
+end
