+++ app/controllers/users/dashboard_controller.rb
@@ -1,3 +1,3 @@
-class Users::DashboardController < ApplicationController
+class Users::DashboardController < Users::BaseController
   def show; end
 end
