+++ app/controllers/users/registrations_controller.rb
@@ -3,7 +3,6 @@
 class Users::RegistrationsController < Devise::RegistrationsController
   before_action :configure_sign_up_params, only: [:create]
   before_action :configure_account_update_params, only: [:update]
-  skip_before_action :require_login
 
   # # GET /resource/sign_up
   # def new
