+++ app/controllers/users/sessions_controller.rb
@@ -2,7 +2,6 @@
 
 class Users::SessionsController < Devise::SessionsController
   before_action :configure_sign_in_params, only: [:create]
-  skip_before_action :require_login
   
   # # GET /resource/sign_in
   # def new
