+++ app/controllers/welcome_controller.rb
@@ -1,5 +1,4 @@
 class WelcomeController < ApplicationController
-  skip_before_action :require_login
 
   def index; end
 end
