+++ app/views/users/dashboard/_pending_invite.html.erb
@@ -1,5 +1,5 @@
 <article class="flex flex-col items-center justify-end flex-wrap border-solid border-4 border-light-blue-500">
-  <%= guest.movie.title %>
-  <%= guest.party.duration%> minutes long<br>
+Movie: <%= link_to guest.movie.title, movie_path(guest.movie.api_id), class: "button bg-blue-200" %><br>
+<%= guest.party.duration%> minutes long<br>
   Starts at: <%= guest.party.time %><br>
 </article>
