+++ app/views/users/dashboard/friends/_pending_friends.html.erb
@@ -1,2 +1,1 @@
 <%= pendfriend.friend.email %>
-<%= pendfriend.status %>
